requirements-basic.txt
requests
flask
flask_sqlalchemy
flask_socketio
langdetect
markdown2
python-dotenv
pymysql
flask-migrate
alembic
flask-security-too
bcrypt>=4.0.1
flask-mail
#spacy
#pyvi
#scikit-learn
#torch
#transformers
#protobuf
#pandas